+++ recipes/database.rb
@@ -1,7 +1,7 @@
 settings = Cacti.settings(node)
 
 if settings['database']['host'] == 'localhost' || settings['database']['host'] == '127.0.0.1'
-  include 'cacti::database_client'
+  include_recipe 'cacti::database_client'
 
   if node['cacti']['mysql_provider'] == 'percona'
     include_recipe 'percona::server'
@@ -27,6 +27,7 @@ if settings['database']['host'] == 'localhost' || settings['database']['host'] =
     ),
     :password => node['mysql']['server_root_password']
   }
+  mysql_cli_opts = "mysql -h #{database_connection[:host]} -u #{database_connection[:username]} -p#{database_connection[:password]} #{settings['database']['name']}"
 
   mysql_database settings['database']['name'] do
     connection database_connection
@@ -34,21 +35,9 @@ if settings['database']['host'] == 'localhost' || settings['database']['host'] =
     notifies :run, 'execute[setup_cacti_database]', :immediately
   end
 
-  cacti_sql_dir = value_for_platform(
-    %w(debian ubuntu) => {
-      'default' => '/usr/share/doc/cacti'
-    },
-    %w(pld) => {
-      'default' => '/usr/share/cacti/sql'
-    },
-    %w(centos fedora redhat) => {
-      'default' => "/usr/share/doc/cacti-#{node['cacti']['version']}"
-    }
-  )
-
   execute 'setup_cacti_database' do
-    cwd cacti_sql_dir
-    command "mysql -h #{database_connection[:host]} -u #{database_connection[:username]} -p#{database_connection[:password]} #{settings['database']['name']} < cacti.sql"
+    cwd node['cacti']['sql_dir']
+    command "#{mysql_cli_opts} < cacti.sql"
     action :nothing
   end
 
@@ -68,34 +57,20 @@ if settings['database']['host'] == 'localhost' || settings['database']['host'] =
     action [:create, :grant]
   end
 
-  # Configure base Cacti settings in database
-  mysql_database 'configure_cacti_database_settings' do
-    connection database_connection
-    database_name settings['database']['name']
-
-    cacti_log_path = value_for_platform(
-      %w(pld) => {
-        'default' => '/var/log/cacti/cacti.log'
-      },
-      %w(centos debian fedora redhat ubuntu) => {
-        'default' => '/usr/share/cacti/log/cacti.log'
-      }
+  template node['cacti']['db_settings_file'] do
+    source 'db-settings.sql.erb'
+    owner node['cacti']['user']
+    group node['cacti']['group']
+    mode 00640
+    variables(
+      :settings => settings
     )
+    notifies :run, 'execute[setup_cacti_settings]', :immediately
+  end
 
-    sql <<-SQL
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_rrdtool","/usr/bin/rrdtool") ON DUPLICATE KEY UPDATE `value`="/usr/bin/rrdtool";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_php_binary","/usr/bin/php") ON DUPLICATE KEY UPDATE `value`="/usr/bin/php";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_snmpwalk","/usr/bin/snmpwalk") ON DUPLICATE KEY UPDATE `value`="/usr/bin/snmpwalk";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_snmpget","/usr/bin/snmpget") ON DUPLICATE KEY UPDATE `value`="/usr/bin/snmpget";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_snmpbulkwalk","/usr/bin/snmpbulkwalk") ON DUPLICATE KEY UPDATE `value`="/usr/bin/snmpbulkwalk";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_snmpgetnext","/usr/bin/snmpgetnext") ON DUPLICATE KEY UPDATE `value`="/usr/bin/snmpgetnext";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_cactilog","#{cacti_log_path}") ON DUPLICATE KEY UPDATE `value`="#{cacti_log_path}";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("snmp_version","net-snmp") ON DUPLICATE KEY UPDATE `value`="net-snmp";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("rrdtool_version","rrd-#{node['cacti']['rrdtool']['version']}.x") ON DUPLICATE KEY UPDATE `value`="rrd-#{node['cacti']['rrdtool']['version']}.x";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_webroot","/usr/share/cacti") ON DUPLICATE KEY UPDATE `value`="/usr/share/cacti";
-      UPDATE `user_auth` SET `password`=md5('#{settings['admin']['password']}'), `must_change_password`="" WHERE `username`='admin';
-      UPDATE `version` SET `cacti`="#{node['cacti']['version']}";
-    SQL
-    action :query
+  execute 'setup_cacti_settings' do
+    cwd node['cacti']['sql_dir']
+    command "#{mysql_cli_opts} < db-settings.sql"
+    action :nothing
   end
 end